move updatesection route to projects.js
