+++ views/projects/showproject.jade
@@ -369,7 +369,7 @@ if (owned)
                                button(type = "button" class = "button button4" class = "NoBtnSection deleteSection" id = "NoButton#{section._id}") No
                                br
                                br
-                form(action='/updateSection',method='post',enctype='application/x-www-form-urlencoded')
+                form(action='/projects/updateSection',method='post',enctype='application/x-www-form-urlencoded')
 
                     input(type='text', name='neumeSectionIds' id = "inputArraySection#{section._id}" hidden)
                     input(type='text', name='SectionID' id = "sectionID#{section._id}" hidden)
